code shortening but still not work
